Separate chip description from chip
