+++ src/chip/mod.rs
@@ -1,5 +1,9 @@
+
+mod chip_description;
+pub use chip_description::ChipDescription;
+
 mod chip;
-pub use chip::{Chip, Link};
+pub use chip::Chip;
 
 #[cfg(test)]
 mod tests;
