+++ src/chip/tests/mod.rs
@@ -2,4 +2,7 @@
 mod test_chip;
 
 #[cfg(test)]
+mod test_chip_description;
+
+#[cfg(test)]
 mod test_circuit;
